pyproject.toml
[project]
name = "km"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu~=1.8.0",
    "httpx~=0.27.0",
    "langchain~=0.2.1",
    "langchain-community~=0.2.1",
    "langchain-core~=0.2.3",
    "langchain-openai~=0.1.7",
    "mcp[cli]~=1.14.0",
    "openai~=1.30.1",
    "pypdf~=4.2.0",
    "python-dotenv~=0.21.1",
]
